Uploaded car.html
I demonstrate object oriented programming in HTML using my car blueprint that I use in some of my other projects. The is only to store data, such as a C struct, with no member functions. To see an example of using this data in a project, visit my other repos, such as ObjectOrientedXMLWebDevelopment and ObjectOrientedJSONWebDevelopment. This data could have been stored in many ways in HTML, such as with tables. This format, using divs, paragraphs, and spans is one the the simplest and most concise. This data can then be parsed, using HTML just like XML or JSON, as a data language ("Markup Language" or "Object Notation").